refactor: check_destination_path in common to be used in contracts, parachain and pallet
